[chore] Move MemberCheckbox file location
